Separando VerificadorTest en VerificadorConOrdenTest y VerificadorSinOrdenTest
