+++ src/main/test/edu/fiuba/algo3/modelo/VerificadorTest/VerificadorConOrdenTest.java
@@ -9,77 +9,46 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 
-public class VerificadorTest {
-
-    @Test
-    public void test01VerificadorSinOrdenRecibe2OpcionesDiferentesYDevuelveFalse(){
-        Verificador verificador = new VerificadorSinOrden();
-        Opciones op1, op2;
-        op1 = new Opciones();
-        op1.agregarOpcion("1");
-        op1.agregarOpcion("2");
-        op1.agregarOpcion("3");
-        op1.agregarOpcion("4");
-        op2 = new Opciones();
-        op2.agregarOpcion("3");
-
-        assertFalse(verificador.verificar(op1, op2));
-
-
-    }
-
-    @Test
-    public void test02VerificadorSinOrdenRecibe2OpcionesIgualesYDevuelveTrue(){
-        Verificador verificador = new VerificadorSinOrden();
-        Opciones op1, op2;
-        op1 = new Opciones();
-        op1.agregarOpcion("1");
-        op1.agregarOpcion("2");
-
-        op2 = new Opciones();
-        op2.agregarOpcion("1");
-        op2.agregarOpcion("2");
-
-        assertTrue(verificador.verificar(op1, op2));
-
-
-    }
+public class VerificadorConOrdenTest {
 
     @Test
     public void test03VerificadorConOrdenRecibe2OpcionesIgualesPeroDeDiferenteOrdenDevuelveFalse(){
         Verificador verificador = new VerificadorConOrden();
         Opciones op1, op2;
+
         op1 = new Opciones();
         op1.agregarOpcion("1");
         op1.agregarOpcion("2");
         op1.agregarOpcion("3");
         op1.agregarOpcion("4");
+
         op2 = new Opciones();
         op1.agregarOpcion("4");
         op1.agregarOpcion("3");
         op1.agregarOpcion("2");
         op1.agregarOpcion("1");
+
         assertFalse(verificador.verificar(op1, op2));
-
-
     }
 
     @Test
     public void test04VerificadorConOrdenRecibe2OpcionesIgualesYDevuelveTrue(){
         Verificador verificador = new VerificadorConOrden();
         Opciones op1, op2;
+
         op1 = new Opciones();
         op1.agregarOpcion("1");
         op1.agregarOpcion("2");
         op1.agregarOpcion("3");
         op1.agregarOpcion("4");
+
         op2 = new Opciones();
         op2.agregarOpcion("1");
         op2.agregarOpcion("2");
         op2.agregarOpcion("3");
         op2.agregarOpcion("4");
-        assertTrue(verificador.verificar(op1, op2));
 
+        assertTrue(verificador.verificar(op1, op2));
     }
 
 }
